docs: add exceptions.py
