+++ docs/methoddocs/exceptions.md
@@ -1,0 +1,6 @@
+# Exceptions
+
+```{eval-rst}
+.. automodule:: ape_alchemy.exceptions
+    :members:
+```
